requirements.txt
certifi==2025.1.31
cloudinary==1.44.0
pillow==11.2.1
python-dotenv==1.1.0
six==1.17.0
urllib3==2.4.0
